Added Users, Profiles, Roles

+++ Controllers/CommentsController.cs
@@ -5,29 +5,42 @@ using System.Web;
 using System.Web.Mvc;
 using System.Data.Entity;
 using MicroSocialPlatform.Models;
+using Microsoft.AspNet.Identity;
 
 namespace MicroSocialPlatform.Controllers
 {
     public class CommentsController : Controller
     {
-        private AppDBContext db = new AppDBContext();
+        private ApplicationDbContext db = new ApplicationDbContext();
         // GET: Comments
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public ActionResult New(Comment comment)
         {
+            comment.UserId = User.Identity.GetUserId();
             var postId = comment.PostId;
             db.Comments.Add(comment);
             db.SaveChanges();
             return RedirectToAction("Show", "Posts", new { id = postId });
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Edit(int id)
         {
             Comment com = db.Comments.Find(id);
-            ViewBag.Comment = com;
-            return View();
+            if (com.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                ViewBag.Comment = com;
+                return View();
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                return RedirectToAction("Index", "Posts");
+            }
         }
         [HttpPut]
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Edit(int id, Post requestCom)
         {
             try
@@ -47,13 +60,23 @@ namespace MicroSocialPlatform.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Delete(int id)
         {
             Comment com = db.Comments.Find(id);
-            var postId = com.PostId;
-            db.Comments.Remove(com);
-            db.SaveChanges();
-            return RedirectToAction("Show", "Posts", new { id = postId });
+
+            if (com.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                var postId = com.PostId;
+                db.Comments.Remove(com);
+                db.SaveChanges();
+                return RedirectToAction("Show", "Posts", new { id = postId });
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                return RedirectToAction("Index", "Posts");
+            }
         }
     }
 }
